requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
black==23.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
coverage==7.10.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.104.1
flake8==6.1.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
isort==5.12.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.19.2
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.12.0
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
orjson==3.11.3
packaging==25.0
pathable==0.4.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pycodestyle==2.11.1
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.1.1
python-multipart==0.0.6
PyYAML==6.0.1
referencing==0.36.2
requests==2.31.0
rfc3339-validator==0.1.4
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
